ui: Associate Wattson estimate per process
This is the process analagous version of the immediately preceding
patchset, which does the attribution on a per thread basis.

Associates Wattson estimate per process based on the user-drawn UI
window. For all processes in the drawn UI window, all processes will be
grouped together across the x-axis (time) and y-axis (CPUs), and the
corresponding Wattson estimate will be summed together.

This feature allows users to attribute Wattson estimates on a process
level, such that extraneous or irrelevant processes can be ignored OR
processes of interest can be selectively chosen for analysis.

Bug: 329507878
Change-Id: I2de520f7e98e4cb86500769ed331f065b647d819
Signed-off-by: Samuel Wu <wusamuel@google.com>