ui: Associate Wattson estimate per package
This is the package analagous version of the previous patches. This
patch does the attribution on a per package basis.

Associates Wattson estimate per package based on the user-drawn UI
window. For all packages in the drawn UI window, all packages will be
grouped together across the x-axis (time) and y-axis (CPUs), and the
corresponding Wattson estimate will be summed together.

This feature allows users to attribute Wattson estimates on a package
level, such that extraneous or irrelevant packages can be ignored OR
packagees of interest can be selectively chosen for analysis.

Bug: 329507878
Change-Id: I07cac98caae141b1d8a3857557897158521c47a9
Signed-off-by: Samuel Wu <wusamuel@google.com>