shared_lib: Add PERFETTO_TE_PROTO_FIELDS param
There are two ways to emit track_event data from the shared_library:
1. The low level ABI:
   * Users can emit protobuf data however they wany.
   * Users have to manually write an if branch a for loop and some lines
     of code for each event.
2. The high level ABI:
   * Users can use a single line macro.
   * There are a limited set of extra params that can be emitted. Custom
     protos are not currently supported.

This commit adds an extra param type to the high level ABI to emit some
types of custom protos.

It's still not possible to have a variable (at runtime) number of
repeated fields: for these use cases the low level ABI is still
required.

However it's going to be possible to build arbitrarily complicated
structures of protos:

```
PERFETTO_TE_PROTO_FIELD_NESTED(5,
  PERFETTO_TE_PROTO_FIELD_CSTR(2, "string"),
  PERFETTO_TE_PROTO_FIELD_VARINT(4, 42)
  PERFETTO_TE_PROTO_FIELD_NESTED(7,
    PERFETTO_TE_PROTO_FIELD_CSTR(5, "string2")
  )
)
```

Another limitation of this mechanism is that there's no (protobuf) type
checking. The user is responsible for picking the appropriate field ids
and choosing the appropriate types.

The main idea behind the implementation of the macro params is to build
temporary arrays of pointer to field types. The arrays are built as
compound literal in C and std::initializer_list in C++.

Bug: 348584514
Change-Id: I5d7b90151db429f7b043411954661e08abc6d631