stdlib: Reorder JOIN to be same order as INDEX
Reorder the sequence in which columns are matched during JOINs to be the
same as the order in which the columns are INDEX-ed in the LUTs. This
ensures that every column in the JOIN ON clause is index-able.

Change in Wattson test runtime (ms) with this patch:
Wattson_dsu_pmu: 14509.02 -> 14995.40
Wattson_suspend: 16073.80 -> 15572.02
Wattson_static_curve_selection: 39375.36 -> 25309.91
Wattson_total_eos_calc: 26227.76 -> 25898.03
Wattson_total_raven_calc: 39600.95 -> 33329.40

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 350065024
Change-Id: Iaaf0345a66e35501bd7704b631b7558d9eeda928
Signed-off-by: Samuel Wu <wusamuel@google.com>