shared_lib: Remove limit of 4 extra params in macro (ABI break)
With this commit, the PERFETTO_TE macro accepts an arbitrary number of
params.

Params are now passed using an array of pointers. In order to create the
array on the stack, the code uses a compound literal in C and a
std::initializer_list is C++.

Extra params earlier were passed to the high-level ABI as a singly
linked list. Instead an array of pointers is used. So the ABI has been
changed to take an array instead.

Change-Id: I677e3d07440cedb28889a4b8bb1865f43c01693a