Read Pixel modem Pigweed events.
Use a (Perfetto) tokenizer to rewrite events for timestamps like we do for power rails and to load the Pigweed database. Then use a parser to look up events in the database.

The modem implementation specifically then also has a tiny key-value format within the rehydrated event. Use this to label tracks for domains.


Bug: 335603768
Change-Id: Iea08b7833f0648d0c48740f4f056ca1a1fff1afe