ui: Remove `frontendLocalState` from `globals.state`
`frontendLocalState` contains information about the visible window,
including resoution and last update time.

This duplicates data stored in `globals.timeline`, and exists purely to
allow the visible window to be shared into permalinks.

Now we have explicit deserialization, we can just update the visible
window directly in `globals.timeline` on permalink load rather than
faffing around with this intermediate state in the store.

A couple of gotchas: The resolution is used by SearchController and
LegacyFlamegraphDetailsPanel, so these areas have been updated.

Whatsmore, controllers rely on updates being made to the store in
order to run, however, now the visible window has been removed from
the state, we no longer update it every time the window moves,
but controllers still need to run based on this visible window.

Seeing as controllers are going away, I put in a placeholder solution
where the timeline just makes empty edits when the timeline moves
which is enough to kick off controllers. This is a massive hack, but
there's not a lot of point fixing how controllers work as we are going
to remove them imminently anyway.

Change-Id: If9959476feb81f32f93346882c7c21703ffeda7e