Fixes recording interface V2 for Windows Desktop.
The is currently an issue with HostOsTarget when WebSocket connection
fails. This causes the reinitialization of the target to |undefined| by
maybeClearTarget. Therefore, when we call targetList to get a list of
the available host target to display we return nothing (an empty array).

The solution here for Windows Desktop (ETW tracing) is to pass through
the chrome tracing path. It is also relevant to pass trough the chrome
target since we pass the necessary information to start ETW through the
same path as we do to start a regular chrome tracing sessions.

The goal is to enable users to start an ETW tracing session directly
through the Perfetto UI, similar to how we initiate Chrome tracing
through the recording page. This will allow users to seamlessly open and
analyze ETW traces within Perfetto.

Change-Id: Ie3b221caeb1eac83a6b2a3d812ff28dd10202ce7