ui: Simplify track lifecycle management using an async generator
This patch replaces the huge ugly state machine that manages the track
lifecycles with a much more elegant approach using an async generator.

Note: This is also applicable for when when we switch to a singleton
track model as we we'll still have to drive the onShow(), onHide()
and onUpdate() functions using the same principles.

Change-Id: I3012bac6ba8281b5ccf9fadc1ae21be6c8ab68c8