metrics: Add recursive hierarchy to estimate breakdown
Add a recursive hierarchy to Wattson estimate breakdown, and add policy
breakdown for the CPU subsystem. An example of what the metric returns
is now like so:

  CPU subsystem
  ├── policy0
  │   ├── cpu0
  │   └── cpu1
  └── policy2
      └── cpu2
    ...
  └── policyN
      ├── cpuX
      ├── cpuY
      └── cpuZ

This patch lists out the estimate for the smallest building blocks of
the CPU subsystem, the individual CPUs and the DSU/SCU, presenting the
info in a tree-like structure that displays the hierarchy of the CPU
subsystem. Additionally, this format is scalable and allows for future
additions of other components (e.g. GPU subsystem, Bluetooth, etc.).

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 340938075
Change-Id: If0fc9cb4cd54b2c09c2c90c895124b452fea8eb7