ui: Change Wattson average power formula
Change the average power formula for Wattson to be
total_energy/total_time instead of total_energy/active_time. While
redundant, this "definition" of average power is the on that power/perf
teams are familiar with using. The formula now is:
  avg_power = total_energy / total_time

The intent of this patch is to reflect the natural intuition that
average power for a process should go up when more threads are
concurrently running, which wasn't the case with the previous iteration
of the code. The relation between average estimated power and total
estimated energy will be easy for the user to establish since the UI
already displays the wall_time/total_time (i.e. time of user drawn
window).

Bug: 350954965
Change-Id: Ibf9c21316ab4f04475fd977b13ce8ac04071f4c7
Signed-off-by: Samuel Wu <wusamuel@google.com>