ui: Use DSU/SCU and dsu_scu for Wattson track
The intent is to illustrate to users that the column is either DSU
(DynamIQ Shared Unit) or SCU (Snoop Control Unit), which are SoC and
trace dependent. The previous iteration with just DSU listed potentially
could be misleading on a trace from a device that doesn't physically
have the hardware.

Since the table directly takes the name from the SQL column, have
dsu_scu as SQL column and thus row name for the "Area Selection" table,
and display DSU/SCU in Wattson tracks.

Bug: 350954965
Change-Id: If24785a02258a788350bab0635b1a220e6fffdb6
Signed-off-by: Samuel Wu <wusamuel@google.com>