ui: Merge `HighPrecisionTimeSpan` and `TimeWindow`.
We currently have two implementations of "high precision" time spans,
one called `HighPrecisionTimeSpan`, and the other called `TimeWindow`.

The primary use-case for this type of span is representing the
current visible window on the timeline.

In order to represent this window, we need to store the start of the
visible window with 64 bit integer precision but also sub-nanosecond
precision to allow inter-nanosecond panning when zoomed close into the
timeline (i.e. where each nanosecond covers several pixels on the
screen).

The duration, however, can just be a number, as we only use this to
convert between pixels which are also represented by numbers.

This CL merges the two time span implementations into one, while also
adjusting the interface so that it works well for it's primary
use-case of representing the current visible window, and improving test
coverage in the process.

This CL also touches the closely related `TimeScale`, which has been
cleaned up and interface adjusted to better fit its use-case.

Change-Id: I1bd59420580796f60362c0fab795238c0ab2f8c5