ui: Remove `windowSpan` and derived properties from globals
The `windowSpan` stores the width of the timeline in pixels. This value
is used by various entities for rendering and resolution calculations.

Currently, this value is written to globals every render cycle from
`ViewerPage`, as this is the mithril component that renders the DOM
elements that make up the timeline. Other components that need
`windowSpan` then reach into globals to access it.

As it turns out, the only components that actually need `windowSpan`
and its derived properties are directly invoked from within
`ViewerPage` (i.e. tracks, panels, etc). Thus, the round-trip through
`globals` is pointless, error prone, and breaks encapsulation.

> Note: The exception is `SearchController` which needs to know the
resolution to load the search summary data. However, this controller
is going to be removed soon, and so this summary data should be loaded
by the panel that actually displays the data - i.e. `TickmarkPanel` -
which is invoked from `ViewerPage`. For now, this CL just instates a
temporary patch in `SearchController` that assumes the timeline is
1000px wide.

In addition, tracks should not depend on `globals` if we want to
support importing them directly from plugins, which is something we
cannot currently do as plugins cannot depend on frontend, which is
where `globals` lives.

This CL removes these properties from globals, and instead just inserts
the window span into any interested parties when invoked from
`ViewerPage` or one of its child components.

For tracks, we introduce a few new objects for passing this data into
various calls - `TrackRenderContext` and `TrackMouseEvent`, for
lifecycle/render calls and mouse event handlers respectively. These
objects include not only the track size but also a precomputed timescale  for convenience. This avoids each track from having to
calculate these values independently (and potentially getting them
wrong), and centralizes this calculation within the TrackPanel,
TrackGroupPanel components.

Change-Id: I9692d8d7ec4766decbadde44b6699dc89c8ed774