Support linux perf files with synthetic events
Also the perf file that surfaced this issue does not record the cpu in
the samples. There was a bug in the importer code that failed to deal
with this properly. This patch also fixes that.

The current importer can not deal with samples with no cpu, so all
these samples will be skipped during import. Bug b/352257666 tracks
a feature request to add support to that.

Change-Id: Ifbe68ee2205b752268187cbe76cf2b48f0603370