ui: improve data loss notification using a popup
This CL adds a popup to flag the user's attention to data loss in the
trace. Too often users miss the fact that there is data loss and are
confused by why the trace appears broken. This makes it much more
obvious that they should seriously look at the data loss page to
check to see if there are problems.

Fixes: 350911539
Change-Id: I1cb390778659c45a48ef224c1197af402de9d74c