ui: Move `labels` from `TrackState` to `TrackDescriptor`
This is part of an effort to move properties from TrackState so that we
can eventually remove this state, and thus trackKeys, entirely.

Testing: Load the test/data/chrome_custom_navigation_trace.gz trace
and ensure the labels are still displayed correctly on the various
track groups that have them.

Change-Id: Ifb40171779cf450f81677539d85c927321ab9bd3