tp: implement support for data descriptors zip files
This CL implements support for zip files which were built in a streaming
fashion and thus just fill zeros in their header for crc/compressed size
and uncompressed size. They put this information in a "data descriptor"
which immediately follows the file.

Add support for this data descriptor by parsing it and backfilling in the header info.

Fixes: https://github.com/google/perfetto/issues/840
Change-Id: Ic9e207788853d7e338390a0dbf6044b5422e6183