metrics: Move Wattson metric to stdlib for reusability
Move Wattson power estimate metric to Perfetto stdlib so that the code
can be re-used.

There are many use-cases for getting estimates from a trace, and the
only difference is the particular time window of interest. Leave the
time window calculation to be metric SQL file specific, but move the
generic estimation and hierarchy breakdown to stdlib.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 352627297
Change-Id: I2d974faa4b65d5b3c28fc6706805a3019ec76ff0
Signed-off-by: Samuel Wu <wusamuel@google.com>