Support standalone logcat parsing
Add support to parse logcat files outside of bug reports. For that we
refactored the parsing code a bit and split it into a tokenization and
a parsing (after sorting) phase. The sorting phase is now needed as we
need to sort the logcat events with events from other trace types.

This cl introduces some logic changes too:

1. When deduplication logcat and dumpstate events we only match each
event once. E.g.
Logcat
ts_1 "Msg 1"

Dumpstate
ts_1 "Msg 1" < This event is not emitted as it is a duplicate
ts_1 "Msg 1" < This event used to be not emitted but is now emitted
               (there is no match for it in the locat)

2. The dedup logic used to ignore the event priority when matching. It
now takes that into account.


Change-Id: Idab950a1e2b1a20451821e2a443335e8f7947409