ui: Add explicit track 'chips'
Plugins can now supply a list of strings to use as track 'chips'
(little tag like objects near the track title), instead of relying on
a couple of well known track tags.

This patch also adds a chip for for thread state and slice tracks that
represent the main thread of a process.

Change-Id: Ic061edeca7110f9c80a98ba6fe8700cf18cf8129