ui: Add track shell button to show track details
Added a new button in the track shell which presents a dropdown when
clicked that shows track information.

This change also plumbs pluginIds through to this details dropdown.

Note: this button is behind a flag as it takes up valuable track shell
real estate, and is mainly only useful for plugin developers.
Change-Id: Id65c2c8470137ca1a5bd8cc8f2cb18c73a149107