[tp] Fix handling the type of NULL columns.
Currently the type of the column for PerfettoSQL tables is inferred from
the storage type. It is a problem for columns which contain only NULLs,
as they default to int storage, which leads to problems downstream when
trying to check the type (e.g. if the user declared the column type as
string and the table is empty, this should not be a type error due to
type of this column being inferred as int).

This patch changes the behaviour of the null columns to be explicitly
marked as "kNull" in the runtime table schema.

Bug: b/353226722
Change-Id: If6a578d6e8dd5269da59b8fc052d825f613ac89e