tp: graph_scan -> graph_aggregating_scan and introduce new graph_scan
This CL changes the existing graph_scan macro to be
graph_aggregating_scan which is a more accurate reflection of what it
is doing. It then introduces a new graph_scan macro which does
something closer to what people naively would expect graph_scan to do.

Also use both operators in the codebase in the relevant places.

Change-Id: If2db99fb7249d9c95d8e3ff26a0e33152fb0e3c6