TP: fail gracefully if a query result row is > 256MB
Today if a query yields a result where a single row is >256MB it
causes a hard-to-debug error of the form:
"RPC framing error, unexpected tag ..."
This is because the act of writing >256MB with protozero creates
a malformed proto message which has fields with sizes of
(real_size % 256MB)*.
This CL detect the issue and replaces the response with a more
graceful message.

* Not quite due to 7-bit varint encoding, but conceptually it's the
same problem.

Bug: 353706393
Change-Id: I8cd90855067a513f765efd9c5dac7fb2c681bb80