UI: try reconnect when websocket connection is lost
On a macbook the act of closing the lid causes websocket connections
to get dropped (even if they are to ws://localhost). This make the
usage of UI+RPC very annoying as the UI crashes upon resume.
Add some code that attempts a reconnection and errors only if the
reconnection failed.

Test: manual: start a UI+RPC session, close&reopen lid. Trace is still usable
Bug: 336596994
Change-Id: I4a4ad4e3fa529f69c6a3eba50c59a4ae81a6e677