ui: Avoid running eslint when no typescript files are changed
When we upload a CL with e.g. only SCSS files, the resultant filtered
list of files passed to eslint is empty which causes eslint to check
all files, which can be annoying.

This CL adds a check to skip the eslint check if the filtered list of
files is empty.

Change-Id: Icf9708a59f9119fc21e65090824435fde1be2b69