ui: add property support to flamegraph and bring back mapping name
This CL adds support to flamegraph widget to display arbitrary
properties on every node. It also adds support for perf samples and
native heap profiles to show mapping names when a node is clicked on
which brings it back to parity with old flamegraph.

Finally, it also adds support for showing file name/line number if those
are non-null.

Fixes: https://github.com/google/perfetto/issues/846
Change-Id: I16610a61d6b1aa6972b63ec5e47a6c55d2a21b51