metrics: Add period_type for Wattson estimation
Add period_type to distinguish between the different possible periods
that a user may want, in addition to extending support to different time
periods Wattson will use in the future.

The output format of Wattson power estimates will all be identical, with
the only difference coming from the "time window" of the trace.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 352627297
Change-Id: Ia86af209e5e3c0b5b0965e6ebec8915d2147bcd1
Signed-off-by: Samuel Wu <wusamuel@google.com>