Extend Perfetto with kprobes tracing
One of the features that the Linux kernel tracing infrastructure offers
is the capability of using kprobes as events.
Kprobe events are very flexible, as it allows user space to trace every
kernel function that is exported, either when they are called (kprobes),
or when they return (kretprobe), and could possibly collect the value of
the function parameters.

This patch enables the most basic kprobe functionality, which include:
- definition of kprobe events for tracing;
- creation of kprobe events in ftrace;
- parsing of kprobe events into a generic kprobe format that generates
  either slices or instants.

This enables Perfetto users to instrument kernel functions, thus being
able to retrieve accurate measurements of function durations and
frequency of function calls.

Test: perfetto_unittests, visual inspection of generated trace
Bug: 359542775
Change-Id: Ia0064be2dc962cd66e70dc64cfc94e2b1093797d
Signed-off-by: Alessio Balsini <balsini@google.com>
Co-developed-by: Daniele Di Proietto <ddiproietto@google.com>