stdlib: Prepend NULL dummy slices per CPU for Wattson
Prepend NULL dummy slices on a per CPU basis for frequency and idle
tables. This is an intermediate step so that each frequency and idle
slice is defined and starts at the same ts.

With this change, we can use interval_intersect/SPAN_JOIN() in lieu of
SPAN_OUTER_JOIN(), which will provide some performance gains.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 350065024
Change-Id: If04460c5b615c6a64298d2825f3e71844bf3f9ae