tp: make NotifyEndOfFile return a base::Status
This CL changes the public API of trace processor to return a
base::Status from the NotifyEndOfFile method. This allows communicating
errors which happen post all the trace file being pushed into TP (e.g.
with gzip streams, we can return errors that the stream was not
sucessfully parsed).

In the future, this method will be made [[nodiscard]] but for now just
keep it a normal method to not break back-compat.

Change-Id: I24ae97465f25c79ae0e340ca01a0b77cefb87a70