stdlib: Fix bug with choosing model name
Fix issue with Wattson stdlib when choosing the model name. The intent
is to choose model or device name, based on if one (or both) pieces of
information are embedded in the metadata, and convert to model.

This will provide generic support based on soc model if that information
is available, extending Wattson support to devices outside of devices
that are explicitly listed.

Also add a new test data trace that does not have an explicitly
supported device name, instead having to rely on the model to determine
if the trace is usable for Wattson.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell
--name-filter '.*wattson.*'
Bug: 356403294
Change-Id: I9d9cac18ed9bb45a1396640c0c126111625228a5
Signed-off-by: Samuel Wu <wusamuel@google.com>