ui: Add track filter using fuzzy name matching
This CL adds a track filter box at the top of the timeline that allows
users to only show tracks whose name match a particular string, using
the same fuzzy-find algo as commands.

This track filtering auto-expands and locks open any groups that
contain matching tracks.

Note: Group names are not searched - only the top-level tracks and
tracks inside groups.

The matching part of the name is displayed in bold, so it's obvious to
the user why a given name matches the given search string. This is
important as the fuzzy find algo is quite permissive, returning a lot
more results than e.g. a simple substring approach.

Change-Id: I77a8e3d602faf597c3b0d08b46a8f180e9f79ef9