tracing_service_impl_unittest: Simplify StopTracingTriggerTimeout
In the test, the tracing session is disabled after a 1ms timeout. If the
execution is slow, the 1ms timeout could happen while the tracing sevice
is still doing the handshake with the producer. In that case, the test
will fail.

This commit simplifies the test and avoids the producer: there are other
tests that check what happens with producers and triggers.

This is useful because a future commit will make TracingServiceImpl more
asynchronous, and this test would have been flaky as a result.

Change-Id: I6f97311c2693bd0ca478b825aec19516eb29f8c7