tracing_service_impl_unittest: Rework timing of internal checks
In a future commit, the service will not observe consumer calls
immediately, but only after a TaskRunner::PostTask.

In tha case we cannot check internals of TracingServiceImpl immediately:
we have to wait.

We can simply wait for the effects to be visible on the producers and
then check the internals.

Change-Id: Ib10ac3949f0906457268fbf122d5ec62eff16d69