ui: Tidy up TrackDecider
Currently, track decider re-runs a lot of the queries that the plugins
have already run to generate the default track list, and attempts to
piece together the URIs of those tracks in the exact same way as the
plugins do.

This is not only redundant, slow, and error prone, but also means it's
impossible to disable plugins in order to avoid showing certain tracks
entirely, as they'll still be added by track decider.

This CL changes the way the bulk of these tracks are added in track
decider to build the default track list instead by querying the list
of available tracks and their metadata.

This CL also sets all the tracks keys to be the same as their URIs,
which makes any operation on tracks that requires track keys (e.g.
pinning a lot simpler).

This CL is the first step towards fixing the track grouping and sorting
logic.

Bug: 357831770
Change-Id: Ie08a532f650e91475aff61b33f16a4263df24a3e