Automatically infer UI thread for each cuj
This deprecates the hardcoded list of ui threads per CUJs in favour of a new detection logic.
In recent Android versions, once a CUJ is emitted, its UIThread will have a slice with format "J<cuj_name>#UIThread".

This slice is used to detect the utid of the cuj main thread.

Note that the old list is still working when this instant event is not available, but the instant event takes priority.

Tests have been added both for the new instant event behaviour, and for the old hardcoded list.

Bug: 337803466
Test: tools/ninja -C out/linux_clang_release && tools/diff_test_trace_processor.py out/linux_clang_release/trace_processor_shell --name-filter AndroidMetrics:android_blocking_calls_cuj_different
Change-Id: If81fe60c6ed5a43ae1ad3889ce04555a43f848c0