tracing_service_impl: Return void from StartTracing()
The return value of StartTracing is ignored everywhere, except in
EnableTracing(). It should not be possible to hit those two error
conditions from EnableTracing(), because we just created the tracing
session in the CONFIGURED state.

This simplifies error handling: callers can assume that if EnableTracing
returns an error, it hasn't allocated a tracing session. It's going to
be helpful for a future change.

Change-Id: Iacec8abfbe152140e3806b66565e895dff577d91