tracing_service_impl: Avoid spurious DLOG
The `DisableTracing` method is called by:
* The consumer directly.
* The `FreeBuffers` method.

In the first case, if a session is CLONED_READ_ONLY, the consumer is
misbehaving, so the log makes sense.

In the second case, it's perfectly ok for a consumer with a
CLONED_READ_ONYL session to call FreeBuffers (which will call
DisableTracing), so the log is not really useful.

Change-Id: I232932725f95067575cd17693312da1f9c60fe59