stdlib: Move deep idle adjustment to be idle exit
Based on empirical data, we found that the CPU is running (active) for
longer than the kernel actually thinks, so we adjust the idle times to
be slightly shorter. This patch changes the adjustment to be on idle
exit instead of idle entry, which can also be thought of as the head
instead of tail end of an active slice.

The adjustment needs to be idle exit, because the next patch in the
series will use idle exit to attribute the exit cost the thread
immediately following the idle exit.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson_thread.*'
Bug: 359959887
Change-Id: I817cafeb2f2305509f79f373fa40d51c628d45ee
Signed-off-by: Samuel Wu <wusamuel@google.com>