[tp] Include query iteration time in the query execution time
When executing the query in trace_processor_shell in the batch mode,
materialise the query first before taking the time measurement.

This ensures that the query iteration cost (which can be non-trivial) is
included in the overall time.

This patch also cleans up metatrace events in PerfettoSqlEngine, ensuring that all Execute* methods are instrumented, as well as standardising the event and arg names.

Bug: b/359182293
Change-Id: I039c74f46b4bb2950e6e642644eb2d044ee20622