traced_perf: Add support to group counters.
This patch enables the sampling of multiple counters at once. When
acquisition happens at the same timestamp it is straightforward to do
arithmetic operation on them as there is no need for normalization of
the counter values involved in the operation around a single timestamp.

For example, consider Instruction Per Cycle (IPC) which is given by the
formula: IPC = (inst_retired) / (cpu_cycle)

If the counters "inst_retired" and "cpu_cycle" are captured at the same
time it is possible to apply the formula for every sample captured.

Otherwise, if capture of these counters is made independently, counter
values at a given timestamp has to be approximated before applying
the formula. This is not precise, straightforward (in SQL) or efficient.

To configure additional counter into a PerfEventConfig you should add
a FollowerEvent to the configuration. It contains a
PerfEvent.Timebase.event and a name as a string.

Here's an example:

perf_event_config {
  timebase {
    frequency: 20000
    counter: HW_CPU_CYCLES
  }
  followers {
    raw_event {
      type: 8
      config: 16
    }
    name: "br_mis_pred"
  }
  followers {
    counter: HW_INSTRUCTIONS
  }
}
Change-Id: I754bbf9a122e01cfccce40050c1e66846c493394