traced_perf: workaround for crash when a uspace thread's pid gets reused by a kthread
We currently are do not precisely track process lifetimes, nor
explictily deal with pid reuse. So the unwinder might see a sample for a
process with a reused pid, while still thinking it's the process that
was first seen under that pid since the data source was started.

Normally this is both infrequent and not a problem since the unwinding
will fail due to invalidated procfs descriptors. However there's a crash
in the specific case of a userspace thread's pid being reused for a
kernel thread, because the code doesn't expect a null userspace register
pointer (specifically, the |sample.regs->sp()| dereference in
UnwindSample).

Work around by detecting this situation. The longer term solution will
be to more accurately track process lifetimes, probably by asking for
PERF_RECORD_EXIT records from the kernel. But there is some subtlety
there because the unwinder might still have inflight samples that are
yet to be unwound.

Bug: 324757089
Change-Id: I86789dc138c7ffc7f5b79daf977db710b4c59dff