ui: Fix circular dep between slice_args.ts <-> sql_table_tab.ts
The cycle looks like this (as reported by rollup).
```
cycle: [
  'ui/tsc/frontend/debug_tracks/debug_tracks.js',
  'ui/tsc/frontend/debug_tracks/slice_track.js',
  'ui/tsc/frontend/debug_tracks/details_tab.js',
  'ui/tsc/frontend/slice_args.js',
  'ui/tsc/frontend/sql_table/tab.js',
  'ui/tsc/frontend/debug_tracks/add_debug_track_menu.js',
  'ui/tsc/frontend/debug_tracks/debug_tracks.js'
],
```

The cycle was broken by making a command for adding a new SQL table tab
where the config is supplied through the command args.

A helper function has been added to help with typing, which only
depends on sql_table_tab.ts for an interface (and thus can use
`import type` to import it), allowing the circular dep to remain broken.

Change-Id: I1a6ccf091d2389573245f8be20b91eed8da24f36