[mac] Emit signpost events on state reset
Instruments/xctrace collects profiles with timestamps relative to the
start of profiling. We can sync these with the perfetto clock by
injecting "point of interest" signpost events into the profile, which
are emitted with their own timestamp and can have another clock's
timestamp as their metadata -- in this case, the boottime, though this
could be extended to other clocks in the future.

Bug: 362665903
Change-Id: I66ff0084ea84ea91437877ec8c9ed22a2bd7391e