ui: tracks: rename methods
A pure renaming CL which uses some clearer names that reflect
the current semantic of the method.
I am aware that some of these things are going to disappear,
but I find so confusing to read the code that it's worth spending
some of my time just for the sanity of the next refactorings.

Change-Id: I6e4140c3f04ae00fd7e9366acc7b3c90b96c6057