Add abbility to create dummy mappings
Some profilers will not emit mapping information so we need to be able
to create dummy mappings. One global dummy mapping (as currently
available) might not be enough if we need to distinguish between
mappings in different processes.

This patch also adds the ability to create dummy frames for profiles
that do not emit an address and just log funciton names.

Change-Id: Id7049050b2157383938620de3cb6ddc26b33c891