Move protolog_viewer_config parsing to TokenizePacket
Protolog viewer config messages don't contain a timestamp this meant that they would be assigned a timestamp and in some cases this timestamp meant that the packet was processed after the protolog message packet that depend on the config, which do provide a timestamp.

This change guarantees that the view config packets are always processed before the protolog message packets to ensure safe and reliable decoding of all the protolog messages.

Bug: 362727234
Change-Id: I65488c4219a6980c265d758f80f606a853f80f00
Test: tools/diff_test_trace_processor.py ./out/linux_clang_release/trace_processor_shell --name-filter "ProtoLog"