ui: Clean up 2D geometry library
The existing 2D geometry library consists of a set of interfaces for
Vectors and Rects, and some naked functions that operate on them. This
is good from a duck typing perspective, as anything that quacks like
a Vector/Rect can be passed in (i.e. mouseEvent or DOMRect).

However, chaining is awkward. One must do:

```ts
vectorAdd(a, vectorSub(b, c))
```

as opposed to (ideally):

```ts
b.sub(c).add(a)
```

It would be nice if we could duck type wherever possible but also allow
method chaining.

This CL introduces a new ethos to the geometry library "accept
interfaces, return objects", which gives us the best of both worlds.

I've also taken the opportunity to add helpful comments, and add
comprehensive unit tests.

Change-Id: Ibe7a5724389cd07afec9ed2204eae9683632f5ce